requirements.txt
requests==2.28.1
selenium==4.3.0
GitPython==3.1.27
Pillow==9.2.0